requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.7.0
drf-spectacular==0.27.2
greenlet==3.1.1
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.20.1
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.1.1
